gitignore raw/ dir & all CSVs in project root
The `raw/` directory can be used as the staging area for the raw CSVs to
be processed. And just in case we're also ignoring any CSVs that might
be left in the repo root.

The main reason to ignore the CSVs is they may contain personal
information that aren't meant to be published to the website.